Break up Input parts. Use ThemedClassName.
