+++ packages/ui/aurora/src/components/Input/index.ts
@@ -3,3 +3,4 @@
 //
 
 export * from './Input';
+export * from './InputMeta';
